feat(api): 完善openapi文档

+++ api/prisma/migrations/20230427134327_init/migration.sql
@@ -15,7 +15,7 @@ CREATE TABLE "user" (
 CREATE TABLE "bill" (
     "id" SERIAL NOT NULL,
     "payType" INTEGER NOT NULL,
-    "amount" INTEGER NOT NULL,
+    "amount" DOUBLE PRECISION NOT NULL,
     "date" TIMESTAMP(3) NOT NULL,
     "typeName" VARCHAR(100) NOT NULL,
     "remark" VARCHAR(100) NOT NULL,
